fix(parser): make png predictor work on bytes rather than samples
The previous implementation applied differences on entire samples rather than bytes. In 8 bit-per-color settings, this meant that only the first color component would be processed. This has now been fixed and is known to work on 8 bpc. Additionally, the TIFF predictor was removed and will remain so until properly implemented.