feat: add typings for PDF dictionaries (and fix everything in between)
We now have TypedDicts for some important PDF dictionaries (catalog and trailer, notably). This means we now depend on typing-extensions for py <3.11 (typing.Required). In between: a) The internal params of CryptFetchFilter are now prefixed with '_'; and b) The check for indirect references in the tokenizer has been updated to hopefully short-circuit and be a bit faster.