chore(parser): add measure for avoiding circular reference loops
If pdfnaut finds a potential circular reference loop while parsing an object, it will now hard fail.

Ideally, if strict is False, we would attempt to guess at this point what was meant. For example, if parsing
a stream's length causes a reference, we would fall back to logic for finding the stream's likely length.