feat(document): create foundations for page assembly
Via the new PageList object and the "pages" attribute, PdfDocument is now capable of inserting
or appending pages to a PDF document. Users can insert pages from within the document or from
another document. The pages are then copied and its references are renumbered (if necessary)
before insertion into the page tree.

For this task, the "Page" object now stores its indirect reference within the document. The object
repr for "Page" has also been modified to only show the page's media box and rotation.

Tests for page insertion and appending are also included in this commit.