feat: drop support for Python 3.9
Stop testing on and, by extension, drop support for Python 3.9. The supported versions now
are 3.10 to 3.14 (both included).

As part of this, move to the natively supported union syntax in dictmodels. In 3.9, 'T | None' would
require a future import. Implementing it as a dictmodel type in this form would be unfeasible.