feat(writer): rewrite serializer to use BytesIO
Performing a bunch of bytes.join and concatenations with immutable objects can get
expensive quite fast, so instead, use a mutable BytesIO buffer for writing the PDF
contents. Some tests showed a 40-50% decrease in time taken to save a document
(test done is reading and immediately saving the document so the time in general
use might be even lower).

Allow the dictmodel decorator to be used without parentheses.

Remove the 'objects' attribute in PdfSerializer as it was not used by the class and
the user can simply create their own mapping of object references to objects.